backend/requirements.txt
boto3==1.35.74
cryptography==44.0.2
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-stubs-ext==5.1.1
drf-spectacular==0.28.0
fsspec==2025.5.1
gunicorn==23.0.0
jwcrypto==1.5.6
kubernetes==32.0.1
psutil==6.1.0
psycopg==3.2.3
pydantic-settings==2.9.1
PyJWT==2.10.1
requests==2.32.3
rustworkx==0.15.1
s3fs==2025.5.1
splink==4.0.9
# Azure Blob Storage support
azure-storage-blob>=12.0.0
azure-identity>=1.0.0
adlfs>=2023.1.0
